requirements.txt
flask
flask-cors
pymongo
gunicorn
python-dotenv
Flask-Mail